Move install function to lib/install.rs
