+++ src/install.rs
@@ -110,8 +110,28 @@ fn install_minify() -> Result<(), Box<dyn std::error::Error>> {
     }
 }
 
-/// Install pandoc and minify.
-pub fn install_dependencies() -> Result<(), Box<dyn std::error::Error>> {
+use os_info;
+/// Install required dependencies.
+pub fn install() {
+    let info = os_info::get();
+    let os_type = info.os_type();
+    let arch = info.architecture();
+
+    let system = (os_type, arch);
+    match system {
+        (os_info::Type::Ubuntu, Some("x86_64")) | (os_info::Type::Debian, Some("x86_64")) => {
+            install_dependencies().expect("cannot install pandoc");
+        }
+        _ => {
+            println!(
+                "The current system (OS= {:?}, architecture= {:?}) is not supported",
+                os_type, arch
+            );
+        }
+    }
+}
+
+fn install_dependencies() -> Result<(), Box<dyn std::error::Error>> {
     match install_pandoc() {
         Ok(_) => (),
         Err(e) => {
